requirements.txt
langchain
pinecone-client
PyPDF2
pypdf
qdrant_client
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
# uncomment to use huggingface llms
huggingface-hub

# uncomment to use instructor embeddings
InstructorEmbedding
sentence-transformers